Move cache etc. modules under sync module
